Develop

+++ db/migrate/20230604073029_create_posts.rb
@@ -4,8 +4,8 @@ class CreatePosts < ActiveRecord::Migration[6.1]
       t.integer :user_id, null: false
       t.string :location_name, null: false
       t.string :address, null: false
-      t.decimal :latitude, null: false #緯度
-      t.decimal :longitude, null: false #経度
+      t.float :latitude, null: false #緯度
+      t.float :longitude, null: false #経度
       t.string :memo, null: false
       t.integer :prefecture, null: false
       t.integer :is_released, null: false, default: 0 #公開・非公開ステータス
